change autogen logic to use dyn.load path of R script
The autogenerated R script now uses its own path to adjust the path of
the shared library. This means the shared library must sit next to the
autogen'd R script.

This was assumed before as well, but for some reason before the code
worked with the paths just referring to the file names (i.e. as
relative from the autogen R source, and *not* the R interpreter path).